feat(dependabot): Adding GitHub Action updates
In this GitHub Actions workflow, Dependabot is configured to perform version updates for two package ecosystems: "github-actions" and "docker". Here's a breakdown of the configuration with comments:

For the "github-actions" package ecosystem:

The directory option is set to "/" to specify that the package manifests are located at the root directory.
The schedule option is set to run the updates on a weekly basis.
The ignore option is used to exclude dependencies with names containing "github-script" from being updated.
For the "docker" package ecosystem:

The directory option is set to "/" to specify that the package manifests (Dockerfile) are located at the root directory.
The schedule option is set to run the updates on a weekly basis.